+++ app/helpers/project_helper.rb
@@ -51,26 +51,6 @@ module ProjectHelper
     tag.to_s.html_safe
   end
 
-  def trust_modification_order_tag(project, current_user)
-    task = trust_modification_order_task(project, current_user)
-    task_state = tags_for_states.fetch(task.status, tags_for_states[:unknown])
-
-    govuk_tag(
-      text: task_state[:text],
-      colour: task_state[:colour]
-    )
-  end
-
-  def direction_to_transfer_tag(project, current_user)
-    task = direction_to_transfer_task(project, current_user)
-    task_state = tags_for_states.fetch(task.status, tags_for_states[:unknown])
-
-    govuk_tag(
-      text: task_state[:text],
-      colour: task_state[:colour]
-    )
-  end
-
   def address_markup(address)
     tag.address address.compact_blank.join("<br/>").html_safe, class: %w[govuk-address]
   end
@@ -89,38 +69,5 @@ module ProjectHelper
     return govuk_tag(text: "Unconfirmed", colour: "grey") if project.tasks_data.academy_details_name.nil?
 
     project.tasks_data.academy_details_name
-  end
-
-  private def trust_modification_order_task(project, current_user)
-    Conversion::Task::TrustModificationOrderTaskForm.new(project.tasks_data, current_user)
-  end
-
-  private def direction_to_transfer_task(project, current_user)
-    Conversion::Task::DirectionToTransferTaskForm.new(project.tasks_data, current_user)
-  end
-
-  private def tags_for_states
-    {
-      not_started: {
-        text: "unconfirmed",
-        colour: "blue"
-      },
-      in_progress: {
-        text: "unconfirmed",
-        colour: "blue"
-      },
-      completed: {
-        text: "needed",
-        colour: "turquoise"
-      },
-      not_applicable: {
-        text: "not applicable",
-        colour: "grey"
-      },
-      unknown: {
-        text: "unknown",
-        colour: "grey"
-      }
-    }
   end
 end