+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- The opening view no longer shows the status of the direction to transfer or
+  trust modification order tasks.
+
 ### Fixed
 
 ## [Release 30][release-30]